regions rest controller

+++ src/main/java/com/tiger8bit/vapeshop/repository/CountryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CountryRepository extends CrudRepository<Country, Long> {
+public interface CountryRepository extends CrudRepository<Country, Integer> {
 
 }
